Add user and device contexts
